+++ src/sequence.cc
@@ -1,12 +1,12 @@
 #include <sequence/sequence.h>
 #include <cmath>
-Sequence::Sequence(SequenceType type, int first_n, int d_q)
+Sequence::Sequence(SequenceType type, float first_n, float d_q)
 {
 	this->type = type;
 	this->first_n = first_n;
 	this->d_q = d_q;
 }
-int Sequence::progression_n(int n)
+float Sequence::progression_n(int n)
 {
 	switch (type)
 	{
